Add a makefile which builds by default and also has a clean target.
The clean function is justified because lack of an idiomatic way to
find and execute a clean target may have contributed to the error that that
previous commit addressed.

The use of a Makefile is justified by an idiomatic way to find and execute the clean function.

The use of the git ls-files is justified beccause of the need to exclude the one
source controlled file in package/lambda from cleanup.