build-and-test.sh: handle non-interactive environments gracefully.
Previously, the first docker run would fail with a short error message:

the input device is not a TTY

Now, if we are running in an automated environment like Jenkins or Circle CI
the interactive option is removed. It is still available, just in case
there is sometimes a reason to run the build interactively.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>